mysite/requirements.txt
asgiref==3.6.0
backports.zoneinfo==0.2.1
Django==4.2
django-taggit==3.0.0
importlib-metadata==6.6.0
Markdown==3.4.3
Pillow==9.5.0
pkg_resources==0.0.0
psycopg2-binary==2.9.3
sqlparse==0.4.4
zipp==3.15.0
